refactor scUtils, testDeployFailureHelper, fix dependencies paths
