+++ test/03_deploy_fail_wrong_worker.test.js
@@ -14,7 +14,7 @@ import {
 } from "../contractLoader";
 import EthCrypto from "eth-crypto";
 import EventEmitter from "eventemitter3";
-import * as constants from "./testConstants";
+import * as constants from "../testConstants";
 
 function sleep(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
